Add files via upload
This is a simple PACMAN GAME. 
I created this using Pygame.
Although I ve worked with Pygame but this was a full fledged Pygame Project which I created. 
Approach: I created a player, added enemies. A player is supposed to catch or eat the coins and the enemy intervenes to stop him. I have created a wall too (wall part I read the tutorials by GS_BLR as I'd to work with keyboard stuff. And finally .png image of the maze was attached to the black coloured GUI.